+++ ntex/src/ws/sink.rs
@@ -1,6 +1,6 @@
 use std::{future::Future, rc::Rc};
 
-use crate::framed::State;
+use crate::framed::{OnDisconnect, State};
 use crate::ws;
 
 pub struct WsSink(Rc<WsSinkInner>);
@@ -15,6 +15,7 @@ impl WsSink {
         Self(Rc::new(WsSinkInner { state, codec }))
     }
 
+    /// Endcode and send message to the peer.
     pub fn send(
         &self,
         item: ws::Message,
@@ -25,5 +26,10 @@ impl WsSink {
             inner.state.write().encode(item, &inner.codec)?;
             Ok(())
         }
+    }
+
+    /// Notify when connection get disconnected
+    pub fn on_disconnect(&self) -> OnDisconnect {
+        self.0.state.on_disconnect()
     }
 }